Palindrome number
Palindrome Number
Given a number N, you need to check whether the given number is a Palindrome or not. A number is said to be Palindrome when it reads the same from backward as forward. Check whether input number N is a palindrome or not. Print "true" if it is a palindrome or "false" otherwise

Sample Input:
5

Sample Output:
true

Sample Input:
121

Sample Output:
true

Common Mistakes:
Make sure you are printing true and false in smallcase only.