Pairs
Given an array of N integers, count the total pairs of integers that have a difference of K.

Input Format
1st line contains N & K (integers).

2nd line contains N numbers of the array. All the N numbers are distinct.

Output Format
One integer - the number of pairs of numbers that have a difference K.

Sample Input:
4 1  
32 0 63 8
Sample Output:
0

Sample Input :
5 2
1 5 3 4 2  
Sample Output :
3

Explanation:
Here we have taken 5 numbers and we want pairs of numbers with a difference of 2. The given array is 1 5 3 4 2.

Pairs having a difference of 2 are: (1,3) (2,4) (3,5) Therefore we got output as 3.