A contest
Contestant who earns a score equal to or greater than the score of the k-th place finisher will advance to the next round, as long as the contestant earns a positive score.

A total of n participants took part in the contest (n≥k), and you already know their scores. Calculate how many participants will advance to the next round.

Input
The first line of the input contains two integers n and k separated by a single space.

The second line contains n space-separated integers a1, a2, ..., an, where ai is the score earned by the participant who got the i-th place. The given sequence is non-increasing, i.e., for all i from 1 to n - 1 the following condition is fulfilled: a(i) ≥ a(i + 1).

Output
Output the number of participants who advance to the next round.

Examples
Input
8 5
10 9 8 7 7 7 5 5
Output
6

Input
4 2
0 0 0 0
Output
0

Explanation
In the first example the participant on the 5th place earned 7 points. As the participant on the 6th place also earned 7 points, there are 6 advancers.

In the second example nobody got a positive score.

Common Mistakes
Remember, k is the position. To get the Kth-place contestant's score you will need to use a[k-1]